03-intro-to-RAG/03.01-basic-RAG/pyproject.toml
[project]
name = "langchain-course"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.32",
    "langchain-pinecone>=0.2.11",
    "langchainhub>=0.1.21",
    "python-dotenv>=1.1.1",
]
